Removed `CaseloadClient` and moved method to `PrisonApiClient`
