+++ server/data/prisonApiClient.ts
@@ -2,7 +2,7 @@ import RestClient from './restClient'
 import type { ApiConfig } from '../config'
 import config from '../config'
 import { prisonApiPaths } from '../paths'
-import type { SentenceAndOffenceDetails } from '@prison-api'
+import type { Caseload, SentenceAndOffenceDetails } from '@prison-api'
 import type { Prisoner } from '@prisoner-search'
 
 export default class PrisonApiClient {
@@ -10,6 +10,12 @@ export default class PrisonApiClient {
 
   constructor(token: Express.User['token']) {
     this.restClient = new RestClient('prisonApiClient', config.apis.prisonApi as ApiConfig, token)
+  }
+
+  async findCurrentUserCaseloads(): Promise<Array<Caseload>> {
+    return (await this.restClient.get({
+      path: prisonApiPaths.caseloads.currentUser({}),
+    })) as Array<Caseload>
   }
 
   async findSentenceAndOffenceDetails(bookingId: Prisoner['bookingId']): Promise<SentenceAndOffenceDetails> {
